moved `bin` to `src/aequitas_cli` and remade as package
