+++ src/aequitas_cli/utils/img/aequitas_report_header.png
