+++ src/aequitas_cli/utils/img/danger.png
