+++ src/aequitas_cli/utils/report.py
